requirements.txt
Django
requests
PyMuPDF
pdfminer.six
openai
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
celery
celery[redis]
redis
django-redis
pyalex
openai
django-cors-headers